fix: removes extra \n when using `getInput`
*description of bug.*
there was an issue with a multiline test i made. the test wasn't
respecting the breaks in the source input file. i am using `readLine()`
to get lines from `stdin`, so i add a linebreak character('\n') when I
add the line to the working `text` variable.

the issue with this is that i was checking the line length before adding
it to the working `text` variable, and since i removed the line break
(when using `readLine()`) the line was empty, and would not add it. i
generally check if something is empty before adding it to an array ..
not sure what i'm trying to fix but feels like i'm being too smart for
my own good here ...

so i fixed it by adding the line break back to the line immediately and
not adding it when adding it to the `text` line, so now the correct
number of lines are made in the input on multiline tests.

*changes*
- cleans up code, adds more comments, moves some stuff into inline
  functions
- adds more `options`, such as ignoring tests, ignoring groups, only run
  tests, only run groups, and output results for all tests (even
  passing) too.
- adds `termcolors` library so the output can be colorful.